Prepared gb-items-source package
