+++ packages/gb-items-source/package.json
@@ -1,6 +1,6 @@
 {
   "name": "gb-items-source",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "type": "module",
   "main": "./dist/index.js",
