+++ packages/gb-items-source/src/maintenance/types.ts
@@ -1,0 +1,3 @@
+import { type Repositories } from '@/types';
+
+export type MaintenanceTask = (repositories: Repositories) => Promise<void>;
