+++ packages/gb-items-source/src/queries/frameGroups.ts
@@ -6,8 +6,8 @@ import {
   type UpdateFrameGroupsParams,
 } from 'gb-printer-schemas';
 import z from 'zod';
-import { getAddTotal } from '@/workers/itemsIndexedDbWorker/queries/helpers/generic';
-import { type ItemsSourceInternal } from '@/workers/itemsIndexedDbWorker/types';
+import { getAddTotal } from '@/queries/helpers/generic';
+import { type ItemsSourceInternal } from '@/types';
 
 
 export async function getFrameGroups(this: ItemsSourceInternal): Promise<ItemsSourceTotalResponse<FrameGroup>> {
