+++ packages/gb-items-source/src/queries/helpers/applyImageTotals.ts
