+++ packages/gb-items-source/src/queries/helpers/resolveOwnership.ts
