+++ packages/gb-items-source/src/queries/helpers/runMaintenance.ts
@@ -1,0 +1,6 @@
+import { startMaintenanceTasks } from '@/maintenance';
+import { type ItemsSourceInternal } from '@/types';
+
+export async function runMaintenance(this: ItemsSourceInternal) {
+  await startMaintenanceTasks(this.repositories);
+}
