+++ packages/gb-printer-schemas/package.json
@@ -1,6 +1,6 @@
 {
   "name": "gb-printer-schemas",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "private": true,
   "type": "module",
   "main": "./dist/index.js",
