+++ packages/gb-printer-web/src/components/Overlays/QueryTool/methods.ts
@@ -1,3 +1,4 @@
+import { ItemsSource } from 'gb-items-source';
 import {
   GetGroupItemsByGroupIdParamsSchema,
   GetHashesByGroupIdParamsSchema,
@@ -25,7 +26,6 @@ import {
   DeleteBinaryItemsByHashesParamsSchema,
 } from 'gb-printer-schemas';
 import z from 'zod';
-import { ItemsSource } from '@/workers/itemsIndexedDbWorker/types';
 
 export type MethodName = keyof Omit<ItemsSource, 'init' | 'runMaintenance'>;
