--- packages/gb-printer-web/src/workers/itemsIndexedDbWorker/maintenance/types.ts
@@ -1,3 +1,0 @@
-import { Repositories } from '@/workers/itemsIndexedDbWorker/repository/entities';
-
-export type MaintenanceTask = (repositories: Repositories) => Promise<void>;
