--- packages/gb-printer-web/src/workers/itemsIndexedDbWorker/queries/helpers/runMaintenance.ts
@@ -1,6 +1,0 @@
-import { startMaintenanceTasks } from '@/workers/itemsIndexedDbWorker/maintenance';
-import { type ItemsSourceInternal } from '@/workers/itemsIndexedDbWorker/types';
-
-export async function runMaintenance(this: ItemsSourceInternal) {
-  await startMaintenanceTasks(this.repositories);
-}
